Move popCount() into new fontTools.misc.intTools
